Menu and Instructions moved from menu.py to interface.py
