+++ hangman/game.py
@@ -1,4 +1,5 @@
 import random
+import os
 from hangman.words import words
 
 class Game:
@@ -24,6 +25,7 @@ class Game:
 
     def start(self):
         while not self.guessed and self.player.health.gethealth() > 0:
+            os.system("cls")
             self.interface.clear()
             self.interface.write_hint(self.hint, self.theme)
             self.interface.write_health(self.player.health.gethealth())
